Fix for 1753: keyboard interaction with meter buttons broken
There are two problems with using Enter to press the meter drop down buttons:
1. It only works the first time.
2. If auto-select auto is on, then audio is auto-selected.

This was caused by commit a0eafd6. The problem with this commit is that meters and sliders are also derived from wxPanel, and therefore affected by the fix.

To fix this, make the above fix specific to the Lyrics panel.