Bug 2080: Using WASAPI, after recording, playback can fail
The bug:
1. Set Audio host to WASAPI
2. choose recording and playback devices which have different default sample rates in shared mode.
3. Set overdub, and playthrough to off.
4. make a short recording.
5. playback. Fails with error dialog.

Cause of bug:
In AudioIO::GetBestRate, the last returned sample rate is cached. If the function is called again, with the same requested sample rate, then the cached value is returned. So in the above steps, when GetBestRate is called for playback, the sample rate of the recording device is returned.

This caching, along with the caching of the supported sample rates with introduced by Al Dimond, back in 2009.

Fix:
Given that the supported sample rates are cached, then I think it's better just to remove the caching of the last returned value. This simplifies the code, rather than adding more code to make the cached last returned sample rate take into account the values of the playing and capturing arguments.